Installed authentification and file gestion
Installed mongoose-unique-validator, bcrypt and jsonwebtoken for user registration and login.

Created files in "routes", "controllers" and "models" for the users registration and authentification.

Installed multer for file gestion

Udpated the delete function to delete the item file

Created auth middleware

Updated "routes/books" with multer and authentification

Installed sharp and resized images with width and height